+++ src/hooks/useTable.test.tsx
@@ -3,7 +3,14 @@ import { getTemplateSrv } from '@grafana/runtime';
 import { renderHook } from '@testing-library/react';
 
 import { ACTIONS_COLUMN_ID } from '@/constants';
-import { CellAggregation, ColumnEditorType, ColumnFilterMode, ColumnFilterType, EditPermissionMode } from '@/types';
+import {
+  CellAggregation,
+  ColumnEditorType,
+  ColumnFilterMode,
+  ColumnFilterType,
+  ColumnPinDirection,
+  EditPermissionMode,
+} from '@/types';
 import {
   columnFilter,
   createColumnAppearanceConfig,
@@ -656,14 +663,14 @@ describe('useTable', () => {
         source: refId,
         name: 'device',
       },
-      pin: true,
+      pin: ColumnPinDirection.LEFT,
     });
     const valueColumn = createColumnConfig({
       field: {
         source: refId,
         name: 'value',
       },
-      pin: false,
+      pin: ColumnPinDirection.NONE,
     });
 
     const { result } = renderHook(() =>